# api/src/controllers/CouponController.ts
import { Controller, Route, Post, Body } from "tsoa";

interface ValidateCouponBody {
  code: string;
  subtotal: number;
}

interface ValidateCouponResponse {
  valid: boolean;
  discount: number;
}

@Route("coupons")
export class CouponController extends Controller {

  @Post("validate")
  public async validate(@Body() body: ValidateCouponBody): Promise<ValidateCouponResponse> {
    return {
      valid: body.code === "IKFOME10",
      discount: body.code === "IKFOME10" ? 10 : 0
    };
  }
}